requirements.txt
torch==2.1.1
xformers==0.0.23
scipy==1.9.3
torchvision
addict
future
lmdb
einops
timm
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
tqdm
loralib
yapf
tacoreader
rasterio
matplotlib
lpips
wandb
tensorboard
pyiqa